Moved user token generation to utils, fixed typos in queries
